Move dimsiss button tests to event form screen
